[frontend] move menu item to one contained button (#10985)
