+++ opencti-platform/opencti-front/src/private/components/workspaces/workspaceHeader/WorkspaceHeader.tsx
@@ -1,11 +1,11 @@
-import React from 'react';
+import React, { useState } from 'react';
 import fileDownload from 'js-file-download';
 import { graphql } from 'react-relay';
 import Typography from '@mui/material/Typography';
 import Tooltip from '@mui/material/Tooltip';
 import { Dashboard_workspace$data } from '@components/workspaces/dashboards/__generated__/Dashboard_workspace.graphql';
 import handleExportJson from 'src/private/components/workspaces/workspaceExportHandler';
-import { fetchQuery } from 'src/relay/environment';
+import { fetchQuery, QueryRenderer } from 'src/relay/environment';
 import Security from 'src/utils/Security';
 import { nowUTC } from 'src/utils/Time';
 import { EXPLORE_EXUPDATE } from 'src/utils/hooks/useGranted';
@@ -17,6 +17,10 @@ import { WorkspaceHeaderToStixReportBundleQuery$data } from '@components/workspa
 import { InvestigationGraph_fragment$data } from '@components/workspaces/investigations/__generated__/InvestigationGraph_fragment.graphql';
 import WorkspaceKebabMenu from '@components/workspaces/WorkspaceKebabMenu';
 import WorkspaceHeaderTagManager from '@components/workspaces/workspaceHeader/WorkspaceHeaderTagManager';
+import Button from '@mui/material/Button';
+import { workspaceEditionQuery } from '@components/workspaces/WorkspacePopover';
+import WorkspaceEditionContainer from '@components/workspaces/WorkspaceEditionContainer';
+import { useFormatter } from '../../../../components/i18n';
 
 const workspaceHeaderToStixReportBundleQuery = graphql`
   query WorkspaceHeaderToStixReportBundleQuery($id: String!) {
@@ -39,12 +43,17 @@ type WorkspaceHeaderProps = {
   handleAddWidget?: () => void;
 };
 
+interface EditionQueryRendererProps {
+  workspace: Dashboard_workspace$data | InvestigationGraph_fragment$data
+}
+
 const WorkspaceHeader = ({
   workspace,
   variant,
   adjust = () => {},
   handleAddWidget = () => {},
 }: WorkspaceHeaderProps) => {
+  const { t_i18n } = useFormatter();
   const { canEdit } = useGetCurrentUserAccessRight(workspace.currentUserAccessRight);
 
   const handleExportDashboard = () => handleExportJson(workspace);
@@ -61,44 +70,79 @@ const WorkspaceHeader = ({
       });
   };
 
+  const [displayEdit, setDisplayEdit] = useState(false);
+  const handleCloseEdit = () => setDisplayEdit(false);
+  const handleOpenEdit = () => {
+    setDisplayEdit(true);
+  };
+
   return (
-    <div style={{ margin: variant === 'dashboard' ? '0 20px' : 0, display: 'flex', justifyContent: 'space-between' }}>
-      <div style={{ display: 'flex', alignItems: 'center' }}>
-        <Tooltip title={workspace.name}>
-          <Typography
-            variant="h1"
-            sx={{ margin: 0 }}
-            style={{ marginRight: canEdit ? 0 : 10 }}
-          >
-            {truncate(workspace.name, 40)}
-          </Typography>
-        </Tooltip>
-        <WorkspaceHeaderTagManager
-          tags={workspace.tags ?? []}
-          workspaceId={workspace.id}
-          canEdit={canEdit}
-        />
-      </div>
-      <div style={{ display: 'flex' }}>
-        <ExportButtons
-          domElementId="container"
-          name={workspace.name}
-          type={workspace.type}
-          adjust={adjust}
-          handleDownloadAsStixReport={handleDownloadAsStixReport}
-          handleExportDashboard={handleExportDashboard}
-        />
-        <WorkspaceKebabMenu workspace={workspace} />
-        {variant === 'dashboard' && (<>
-          <Security
-            needs={[EXPLORE_EXUPDATE]}
-            hasAccess={canEdit}
-          >
-            <WorkspaceWidgetConfig onComplete={handleAddWidget} workspace={workspace} />
+    <>
+      <div style={{ margin: variant === 'dashboard' ? '0 20px' : 0, display: 'flex', justifyContent: 'space-between' }}>
+        <div style={{ display: 'flex', alignItems: 'center' }}>
+          <Tooltip title={workspace.name}>
+            <Typography
+              variant="h1"
+              sx={{ margin: 0 }}
+              style={{ marginRight: canEdit ? 0 : 10 }}
+            >
+              {truncate(workspace.name, 40)}
+            </Typography>
+          </Tooltip>
+          <WorkspaceHeaderTagManager
+            tags={workspace.tags ?? []}
+            workspaceId={workspace.id}
+            canEdit={canEdit}
+          />
+        </div>
+        <div style={{ display: 'flex' }}>
+          <ExportButtons
+            domElementId="container"
+            name={workspace.name}
+            type={workspace.type}
+            adjust={adjust}
+            handleDownloadAsStixReport={handleDownloadAsStixReport}
+            handleExportDashboard={handleExportDashboard}
+          />
+          <WorkspaceKebabMenu workspace={workspace} />
+          {variant === 'dashboard' && (<>
+            <Security
+              needs={[EXPLORE_EXUPDATE]}
+              hasAccess={canEdit}
+            >
+              <WorkspaceWidgetConfig onComplete={handleAddWidget} workspace={workspace} />
+            </Security>
+          </>)}
+          <Security needs={[EXPLORE_EXUPDATE]} hasAccess={canEdit}>
+            <Button
+              variant='contained'
+              disableElevation
+              sx={{ marginLeft: 1 }}
+              onClick={handleOpenEdit}
+            >
+              {t_i18n('Update')}
+            </Button>
           </Security>
-        </>)}
+        </div>
       </div>
-    </div>
+      <QueryRenderer
+        query={workspaceEditionQuery}
+        variables={{ id: workspace.id }}
+        render={({ props: editionProps }: { props: EditionQueryRendererProps }) => {
+          if (!editionProps) {
+            return <div />;
+          }
+          return (
+            <WorkspaceEditionContainer
+              workspace={editionProps.workspace}
+              handleClose={handleCloseEdit}
+              open={displayEdit}
+              type={workspace.type}
+            />
+          );
+        }}
+      />
+    </>
   );
 };
